pcworks

+++ filemanager.py
@@ -1,4 +1,7 @@
-#!/data/data/com.termux/files/usr/bin/python
+''':'
+python3 $0 $@
+exit
+':'''
 from http.server import BaseHTTPRequestHandler, HTTPServer
 import time
 from urllib.request import urlopen as u
@@ -8,25 +11,42 @@ from json import loads as l
 import os
 import random
 
+def dlistdir(q):
+ return [w for w in os.listdir(q) if w != '...']
 os.chdir('/'.join(argv[0].split('/')[:-1]))
-
-os.system('rm hexit')
-
-a=l(open('listfiles').read())
+os.system('rm exit')
+a=l(open('0/...').read())
 for w in list(a.keys()):
- if a[w] != os.path.getsize(w):
+ if not os.path.exists(w):
+  del(a[w])
+ elif a[w] != os.path.getsize(w):
   del(a[w])
   os.remove(w)
-  print(w)
-open('listfiles','w').write(str(a).replace("'",'"'))
+for w in dlistdir('../0f/'):
+ if os.path.isfile(w):
+  a[w]=os.path.getsize(w)
+open('0/...','w').write(str(a).replace("'",'"'))
 
 q=[w.split() for w in os.popen('ifconfig').read().split('\n')]
-q=[w[1].split(':')[1] for w in q if len(w) and w[0] == 'inet']
+q=[w[1] for w in q if len(w) and w[0] == 'inet']
+q=[w[([int(ww in '1234567890') for ww in w]+[1]).index(1):] for w in q]
+q=[w[:([int(ww not in '1234567890.') for ww in w]+[1]).index(1)] for w in q]
 print('='*10,*enumerate(q),sep='\n')
 if len(argv) > 1:
- port=int(argv[1])
+ argv1=argv[1]
+ if argv1 in [w[:len(argv1)] for w in q]:
+  q=[w for w in q if w[:len(argv1)] == argv1]
+print('='*10,*enumerate(q),sep='\n')
+if len(argv) > 2:
+ port=int(argv[2])
 else:
  port=9000
+if len(q) > 1:
+ e=int(input())
+else:
+ e=0
+q=q[e]
+hostName = q
 hostPort = port
 def exex():
  global hostName
@@ -34,7 +54,7 @@ def exex():
  #print(1234)
  #port=os.system('echo "123" && ./fileserverstarter '+hostName+' '+str(hostPort))
  #print('='*2)
- port=os.popen('./fileport '+hostName+' '+str(hostPort)+' 0').read().split('='*9)[-1][1:]
+ port=os.popen('./fileport.py '+hostName+' '+str(hostPort)+' 0').read().split('='*9)[-1][1:]
  #print('='*3,port)
  return port[:-1]
 class MyServer(BaseHTTPRequestHandler):
@@ -67,15 +87,6 @@ class MyServer(BaseHTTPRequestHandler):
    </head>
    </html>''', "utf-8"))
 
-for w in q:
- os.system('./filemanager '+str(w)+' '+str(port)+' & > log')
-exit()
-try:
- while 1:
-  time.sleep(1)
-except:
- pass
-'''
 st=1
 while st:
  try:
@@ -89,8 +100,7 @@ try:
     myServer.serve_forever()
 except KeyboardInterrupt:
     pass
-'''
-os.system('touch hexit')
-#myServer.server_close()
-#print()
-#print(time.asctime(), "Server Stops - %s:%s" % (hostName, hostPort))
+os.system('touch exit')
+myServer.server_close()
+print()
+print(time.asctime(), "Server Stops - %s:%s" % (hostName, hostPort))
